move Kubernetes helper code into separate file
The goal is to have only "legacy" code in klog.go.
